rename `constants.py` -> `config.py`
