+++ machine/tests/integration/charm_states/ha_proxy.py
@@ -3,7 +3,7 @@ from asyncio import Task, create_task
 import pytest
 from pytest_operator.plugin import OpsTest
 
-from tests.integration.constants import HAPROXY_APPLICATION_NAME
+from tests.integration.config import HAPROXY_APPLICATION_NAME
 from tests.integration.helpers import deploy_if_not_exists
