+++ machine/tests/integration/charm_states/self_signed_certificates.py
@@ -3,7 +3,7 @@ from asyncio import Task, create_task
 import pytest
 from pytest_operator.plugin import OpsTest
 
-from tests.integration.constants import (
+from tests.integration.config import (
     JUJU_FAST_INTERVAL,
     SELF_SIGNED_CERTIFICATES_APPLICATION_NAME,
     SELF_SIGNED_CERTIFICATES_REVISION,
