+++ machine/tests/integration/charm_states/vault_kv_requirer.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import pytest
 from pytest_operator.plugin import OpsTest
 
-from tests.integration.constants import VAULT_KV_REQUIRER_APPLICATION_NAME
+from tests.integration.config import VAULT_KV_REQUIRER_APPLICATION_NAME
 from tests.integration.helpers import deploy_if_not_exists
