+++ machine/tests/integration/test_kv.py
@@ -3,7 +3,7 @@ from asyncio import Task
 import pytest
 from pytest_operator.plugin import OpsTest
 
-from tests.integration.constants import (
+from tests.integration.config import (
     APP_NAME,
     JUJU_FAST_INTERVAL,
     VAULT_KV_REQUIRER_APPLICATION_NAME,
